Rework to allow private repositories.
Add support for storing the git configuration.

Add support for writing out a mapping to the config file.

Add support for passing the auth through to go-scm's Factory function.

Bump to the latest version of status tracker image in use.

Add support in the webhooks, for loading the manifest and configure the drivers on start.